Use KUBERNETES_NODE_NAME
To avoid abbreviations that may be meaningless to others, prefer
KUBERNETES_NODE_NAME to get the name of the node rather than
K8S_NODE_NAME, which we want to deprecate.

Flesh out the docstrings for get_node, get_digest, and get_pod,
prefer the Nublado v3 code path, and be clearer about the purpose
of the fallback. Standardize on reference for the name of the
fully-qualified Docker image reference.