Use urljoin in get_service_url
The URLs returned by get_service_url contained a doubled slash if
<service>_ROUTE began with a slash, causing all of the PyVO
authentication configuration to also be set up with a doubled slash.
PyVO sometimes canonicalizes URLs internally, which will remove the
second slash, and then the authentication configuration didn't match.
This showed up as a failure to authenticate async TAP requests.

Use urljoin to construct the service URL returned by get_service_url,
which avoids this problem. The drawback of this approach is that if
EXTERNAL_INSTANCE_URL contains a path component (not just a hostname
with an optional trailing slash), it must end with a slash or the
last part of the path component will be removed by urljoin following
normal HTML link semantics.