Simplify testing instructions
Don't create a full-blown kernel to test development versions of
lsst.rsp. This is no longer needed since the JupyterLab environment
is independent of the lsst.rsp instance installed in the kernel, so
one doesn't need special measures to avoid breaking the lab startup.

Instead, recommend installing lsst.rsp in your local user environment
when testing, and then uninstalling it when finished testing.